+++ lib/src/core/core.dart
@@ -1,5 +1,6 @@
 /// Core (non-feature) exports.
 library;
 
+export 'app_identifier.dart';
 export 'cancel_token.dart';
 export 'core_manager.dart';